fix(ringbuffer): use native size for find offsets
rbuf_find_bytes() scans up to the ringbuffer byte-count domain, but reported the found offset through an int pointer. On a ringbuffer whose readable data exceeds INT_MAX bytes, a full scan or late match could overflow the signed output counter even though the buffer itself is sized and indexed with size_t.

Current vs potential bug: current ACLK HTTP and WebSocket users keep buffers below INT_MAX, so this is not a known currently exercised production corruption path in those callers. The generic ringbuffer API nevertheless had a deterministic type-width mismatch for large buffers.

Solution: change rbuf_find_bytes() to report offsets through size_t *, update every direct in-repo caller to use size_t local indexes, and adjust WebSocket diagnostics to use size_t formatting. Focused unit coverage now verifies match and no-match offsets.

Functional impact: existing parsing behavior and return codes are unchanged. Offsets still represent bytes from the current tail; the output type now matches the rest of the ringbuffer capacity and byte-count API.

Performance impact: neutral. The scan loop still increments one pointer and one offset per byte; the offset type now uses the native size_t width already used by the surrounding ringbuffer code.

Validation: inspected the staged diff and all direct rbuf_find_bytes() call sites. Worker validation passed whitespace checks, changed object compilation, a local ringbuffer unittest runner, and reviewer quorum after rerunning an incomplete reviewer output; a broader netdata target stopped on an unrelated missing ACLK proto after changed objects compiled.
(cherry picked from commit 17b2c7f24ffd19cd40c8ff1d9670e74be7b08998)
(cherry picked from commit 066d605ba2afd93e70b290948e4c98921181132f)

+++ src/aclk/mqtt_websockets/ws_client.c
@@ -245,14 +245,14 @@ exit_with_error:
 
 #define HTTP_HDR_LINE_CHECK_LIMIT(x)                                                                                   \
     if ((x) >= MAX_HTTP_LINE_LENGTH) {                                                                                 \
-        nd_log(NDLS_DAEMON, NDLP_ERR, "ACLK: HTTP line received is too long. Maximum is %d", MAX_HTTP_LINE_LENGTH);                                  \
+        nd_log(NDLS_DAEMON, NDLP_ERR, "ACLK: HTTP line received is too long. Maximum is %zu", (size_t)MAX_HTTP_LINE_LENGTH);                         \
         return WS_CLIENT_PROTOCOL_ERROR;                                                                               \
     }
 
 int ws_client_parse_handshake_resp(ws_client *client)
 {
     char buf[HTTP_SC_LENGTH];
-    int idx_crlf, idx_sep;
+    size_t idx_crlf, idx_sep;
     char *ptr;
     size_t bytes;
 
@@ -316,13 +316,13 @@ int ws_client_parse_handshake_resp(ws_client *client)
                 nd_log(NDLS_DAEMON, NDLP_ERR, "ACLK: Expected HTTP hdr field key/value separator \": \" before endline in non empty HTTP header line");
                 return WS_CLIENT_PROTOCOL_ERROR;
             }
-            if (idx_crlf == idx_sep + (int)strlen(HTTP_HDR_SEPARATOR)) {
+            if (idx_crlf == idx_sep + strlen(HTTP_HDR_SEPARATOR)) {
                 nd_log(NDLS_DAEMON, NDLP_ERR, "ACLK: HTTP Header value cannot be empty");
                 return WS_CLIENT_PROTOCOL_ERROR;
             }
 
             if (idx_sep > HTTP_HEADER_NAME_MAX_LEN) {
-                nd_log(NDLS_DAEMON, NDLP_ERR, "ACLK: HTTP header too long (%d)", idx_sep);
+                nd_log(NDLS_DAEMON, NDLP_ERR, "ACLK: HTTP header too long (%zu)", idx_sep);
                 return WS_CLIENT_PROTOCOL_ERROR;
             }
